WIP: fixing project structure an breaking down logic in different files
