+++ lib/constants/utils.constants.js
@@ -1,0 +1,8 @@
+const alias = {
+    'model': 'm',
+    'extension': 'e',
+    'path': ['p', 'sequelize-path'],
+    'export': ['export-class']
+}
+
+module.exports = { alias }
