+++ lib/services/index.js
@@ -1,0 +1,6 @@
+const ModelService = require('./model.service')
+const SequelizeService = require('./sequelize.service')
+
+module.exports = {
+    ModelService, SequelizeService
+}
